moved Output class to output.py #1
